+++ routes/UserRoute/GetUser.js
@@ -1,4 +1,4 @@
-const Users = require('../models/User').User;
+const Users = require('../../models/User').User;
 
 const GetUser = async(req, res) => {
     const { id } = req.params;
